+++ render_static/transpilers/__init__.py
@@ -10,9 +10,11 @@ from datetime import date, datetime
 from enum import Enum
 from types import ModuleType
 from typing import (
+    TYPE_CHECKING,
     Any,
     Callable,
     Collection,
+    Dict,
     Generator,
     List,
     Optional,
@@ -25,6 +27,12 @@ from warnings import warn
 from django.apps import apps
 from django.apps.config import AppConfig
 from django.utils.module_loading import import_module, import_string
+
+if TYPE_CHECKING:
+    from render_static.templatetags.render_static import (
+        OverrideNode,  # pragma: no cover
+    )
+
 
 __all__ = [
     'to_js',
@@ -142,6 +150,27 @@ class CodeWriter:
         self.prefix_ = prefix or ''
         self.nl_ = self.nl_ if self.indent_ else ''  # pylint: disable=C0103
 
+    def get_line(self, line: Optional[str]) -> str:
+        """
+        Returns a line with indentation and newline.
+        """
+        return f'{self.prefix_}{self.indent_ * self.level_}{line}{self.nl_}'
+
+    def get_lines(self, lines: Generator[Optional[str], None, None]) -> str:
+        """
+        Returns a string of lines from a generator with the indentation and 
+        newlines added. This is meant to be used in place during overrides,
+        so the first newline has no indents. So it will have the line prefix
+        of {{ default_impl }}.
+        """
+        lines_str = ''
+        for idx, line in enumerate(lines):
+            if idx == 0:
+                lines_str += f'{line}{self.nl_}'
+            else:
+                lines_str += self.get_line(line)
+        return lines_str
+
     def write_line(self, line: Optional[str]) -> None:
         """
         Writes a line to the rendered code file, respecting
@@ -151,8 +180,7 @@ class CodeWriter:
         :return:
         """
         if line is not None:
-            self.rendered_ += f'{self.prefix_}{self.indent_ * self.level_}' \
-                              f'{line}{self.nl_}'
+            self.rendered_ += self.get_line(line)
 
     def indent(self, incr: int = 1) -> None:
         """
@@ -195,6 +223,8 @@ class Transpiler(CodeWriter, metaclass=ABCMeta):
     parents_: List[Union[ModuleType, Type[Any]]]
     target_: ResolvedTranspilerTarget
 
+    overrides_: Dict[str, 'OverrideNode']
+
     @property
     def target(self):
         """The python artifact that is the target to transpile."""
@@ -215,6 +245,7 @@ class Transpiler(CodeWriter, metaclass=ABCMeta):
     def __init__(
         self,
         to_javascript: Union[str, Callable] = to_javascript_,
+        overrides: Optional[Dict[str, 'OverrideNode']] = None,
         **kwargs
     ) -> None:
         super().__init__(**kwargs)
@@ -223,6 +254,7 @@ class Transpiler(CodeWriter, metaclass=ABCMeta):
             if callable(to_javascript)
             else import_string(to_javascript)
         )
+        self.overrides_ = overrides or {}
         self.parents_ = []
         assert callable(self.to_javascript), 'To_javascript is not callable!'
 
@@ -238,7 +270,7 @@ class Transpiler(CodeWriter, metaclass=ABCMeta):
         """
         return True
 
-    def transpile(  # pylint: disable=too-many-branches
+    def transpile(  # pylint: disable=too-many-branches, disable=too-many-statements
             self,
             targets: TranspilerTargets
     ) -> str:
@@ -288,12 +320,31 @@ class Transpiler(CodeWriter, metaclass=ABCMeta):
                     try:
                         target = apps.get_app_config(target)
                     except LookupError:
-                        try:
-                            target = import_string(target)
-                        except ImportError:
-                            # this is needed when there is no __init__ file in
-                            # the same directory
-                            target = import_module(target)
+                        parts = target.split('.')
+                        tries = 0
+                        while True:
+                            try:
+                                tries += 1
+                                target = import_string(
+                                    '.'.join(
+                                        parts[0:
+                                            None if tries == 1
+                                            else -(tries-1)
+                                        ])
+                                )
+                                if tries > 1:
+                                    for attr in parts[-(tries-1):]:
+                                        target = getattr(target, attr)
+                                break
+                            except (ImportError, AttributeError):
+                                if tries == 1:
+                                    try:
+                                        target = import_module('.'.join(parts))
+                                    except (ImportError, ModuleNotFoundError):
+                                        if len(parts) == 1:
+                                            raise
+                                elif tries == len(parts):
+                                    raise
 
             node = _TargetTreeNode(target, self.include_target(target))